Convert all pages to server components
- Created client components for interactive features:
  * HomeContent for product browsing with filters
  * ProductSellForm for product creation
  * ProfileContent, ProfileEditor, UserProfile for profile management
  * AuthLogin, AuthRegister, EmailVerifier, PasswordReset, PasswordResetForm for authentication
- Converted all page components to server components
- Added Suspense boundary for useSearchParams() in verify-email page
- Temporarily disabled Google Fonts due to network issues
- All pages now properly use server components with client components for interactivity
- Build passes successfully with no errors
