+++ app/(auth)/verify-email/page.tsx
@@ -1,17 +1,5 @@
-"use client";
-
-import { useState, useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { useMutation } from "@tanstack/react-query";
-import { EmailVerification } from "@/components/molecules/AuthForm/EmailVerification";
-import { Container } from "@/components/atoms/Container/Container";
-
-// TODO: Replace with actual API call
-const resendVerificationEmail = async (email: string): Promise<void> => {
-  console.log("Resend verification to:", email);
-  // Simulate API call
-  await new Promise((resolve) => setTimeout(resolve, 1000));
-};
+import { Suspense } from "react";
+import { EmailVerifier } from "@/components/organisms/AuthPage/EmailVerifier";
 
 /**
  * Email Verification Page
@@ -19,47 +7,10 @@ const resendVerificationEmail = async (email: string): Promise<void> => {
  * Email verification prompt and resend functionality
  */
 export default function VerifyEmailPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const email = searchParams.get("email") || "";
-  const [success, setSuccess] = useState(false);
-
-  const resendMutation = useMutation({
-    mutationFn: resendVerificationEmail,
-  });
-
-  const handleResend = () => {
-    resendMutation.mutate(email);
-  };
-
-  const handleVerified = () => {
-    // Redirect to home or profile setup
-    router.push("/");
-  };
-
-  // Check for verification token in URL (when user clicks email link)
-  useEffect(() => {
-    const token = searchParams.get("token");
-    if (token) {
-      // TODO: Verify the token with backend
-      console.log("Verifying token:", token);
-      setSuccess(true);
-    }
-  }, [searchParams]);
-
-  const loading = resendMutation.isPending;
-  const error = resendMutation.error ? "Greška pri slanju email-a. Molimo pokušajte ponovo." : undefined;
-
+  // Server component with Suspense boundary for useSearchParams
   return (
-    <Container className="py-8">
-      <EmailVerification
-        email={email}
-        onResend={handleResend}
-        onVerified={handleVerified}
-        loading={loading}
-        success={success}
-        error={error}
-      />
-    </Container>
+    <Suspense fallback={<div className="flex justify-center items-center min-h-screen">Loading...</div>}>
+      <EmailVerifier />
+    </Suspense>
   );
 }
